Fix: 퀴즈 컨트롤러 생성
